requirements.txt
selenium
pandas
lxml
python-dotenv